# Conflicts:
#	globals.py
#	view/new_observation_page.py
#	view/selection_sensor_quantity_page.py
#	view/summary_admin_page.py

+++ model/remote.py
@@ -22,27 +22,34 @@ cursor = None
 thread_active = 0  # is used to know if the program is actively trying to reconnect to the remote db
 disconnect_request = 0  # is used to stop the connect thread from looping
 
-ids_to_modify = ['id_sensor', 'id_system', 'id_data', 'id_observation']
+flag_synchro = False
+
+ids_to_modify = ['id_sensor', 'id_data', 'id_observation']
+tables_to_prepend = ['sensor', 'data', 'observation']
+
 
 def connect_to_remote_db():
     """!
     Tries to connect to the remote database at an interval of 2 seconds. This function is used by other functions in
     an asynchronous manner using a thread, so it doesn't block the main program
     """
-    global db, cursor, thread_active, disconnect_request
+    global db, cursor, thread_active, disconnect_request, flag_synchro
 
     thread_active = 1
     disconnect_request = 0
     while not disconnect_request:
         try:
             db = mysql.connector.connect(
-                host="127.0.0.1",
+                host="localhost",
                 user="root",
                 password="Q3fhllj2",
                 database="prisme@home_ICM"
             )
             cursor = db.cursor()
+            print("Connected to remote database")
+            flag_synchro = True
             synchronise_queries()
+            flag_synchro = False
             thread_active = 0
             break  # Break out of the loop if synchronization is successful
 
@@ -50,10 +57,10 @@ def connect_to_remote_db():
             print("Failed to connect to remote DB. Retrying")
             print(e)
             time.sleep(2)
-    else:
-        print("Not allowed to join")
-        time.sleep(2)
 
+    else:
+        print("Disconnect request")
+        time.sleep(2)
 
 def disconnect_from_remote_db():
     """!
@@ -63,51 +70,55 @@ def disconnect_from_remote_db():
     global db, cursor, disconnect_request
     if db is not None and db.is_connected():
         try:
+            disconnect_request = 1
+            cursor.close()
+            db.close()
             db = None
             cursor = None
-            disconnect_request = 1
             return 1
 
         except Exception as e:
-            print("failed to disconnect from distant db")
+            print("\033[91mfailed to disconnect from distant db\033[0m")
             print(e)
     else:
         print("already disconnected from distant db")
 
 
-def execute_remote_query(query, values=None):
+def execute_remote_query(query, values=None, synchronise=False):
     """!
     Tries to insert into the remote database the query and values passed as arguments
     @param query the query to be inserted
     @param values the values of said query
+    @param synchronise: TODO
     @return 1 if successfully inserted, 0 otherwise
     """
-    global db, cursor
+    global db, cursor, flag_synchro
 
-    if db is not None and db.is_connected():  # checks if connected to the remote DB
+    if (db is not None and db.is_connected()):  # checks if connected to the remote DB
         try:
+            if not synchronise:
+                while flag_synchro:
+                    pass
+
             if values is not None:
-                print("try execute_remote_query : ")
-                print("query", query, "values", values)
+                print("\033[94mExecuted in remote : ", query, "values", values, "\033[0m")
                 cursor.execute(query, values)
             else:
+                print("\033[94mExecuted in remote : ", query, "\033[0m")
                 cursor.execute(query)
+
             db.commit()  # No errors, query inserted in the remote db
             return 1
         except mysql.connector.Error as error:
             # Error inserting the data in distant base
-            # TODO : Matteo return autre chose et ne pas cache la requête ici
-            # Si la requête a eu une erreur ici, elle ne se fera jamais
-            # Cas normalement impossible mais soyons rigoureux
-            print("try execute_remote_query : ")
-            print("query", query, "values", values)
-            print(error)
-            return 0
+            print("\033[91mErreur ", error, "En executant : ", query, "values", values, "\033[0m")
+            raise
     else:
-        print("Distant database not connected")
+        print("\033[91mDistant database not connected req : ", query, "values", values, "\033[0m")
+
         # Create thread to check on the database
+        local.caching = True
         if thread_active == 0:
-            local.caching = True
             connection_thread = threading.Thread(target=connect_to_remote_db)
             connection_thread.start()
         return 0
@@ -122,28 +133,55 @@ def synchronise_queries():
     print("Synchronising data between local and remote databases")
     if db is not None and db.is_connected():
         try:
-            # Fetch all entries from the 'remote_queries' table
-            while local.caching == True:
-                time.sleep(0.5)
+            print("local.caching = ", local.caching)
+            while local.caching:
+                print("Wait for the end of caching")
+            print("Fin while local.caching")
+            try:
+                print("Entering 2nd try sychro")
+                if local.local_db is not None and local.local_db.is_connected():
+                    # Fetch all entries from the 'remote_queries' table
+                    print("Waiting for local.cursor protection")
+                    with local.local_cursor_protection:
+                        print("Executing query SELECT * FROM remote_queries")
+                        local.local_cursor.execute("SELECT * FROM remote_queries")
+                        print("Selected")
+                        queries = local.local_cursor.fetchall()
+                        print("Fetched")
+                        if queries is None:
+                            print("\033[93msynchronise_queries : No queries\033[0m")
+                            return
+                else:
+                    print("\033[93msynchronise_queries : PB DB\033[0m")
+                    local.connect_to_local_db()
+                    synchronise_queries()
+                    return
 
-            local.local_cursor.execute("SELECT * FROM remote_queries")
-            queries = local.local_cursor.fetchall()
+            except Exception as e:
+                print(f"\033[91mError select from remote_queries: {e}\033[0m")
+                if e.errno == 2013:
+                    connect_to_remote_db()
+                    return
+                print("Try to synchronise again")
+                synchronise_queries()
+                return
+
             for query_entry in queries:
-                print("Query: ", query_entry[1])
-                query = query_entry[1]  # Assuming the query is in the first column
-                success = execute_remote_query(query)
+                print("Syncro req : ", query_entry[1])
+                query = query_entry[1]
+                success = execute_remote_query(query,None,True)
                 if success:
-                    # If the query was executed successfully, delete the entry from the local table
-                    local.local_cursor.execute("DELETE FROM remote_queries WHERE query = %s", (query,))
-                    local.local_db.commit()
-                    print("YEEEESSSSS")
+                    if local.local_db is not None and local.local_db.is_connected():
+                        # If the query was executed successfully, delete the entry from the local table
+                        with local.local_cursor_protection:
+                            local.local_cursor.execute("DELETE FROM remote_queries WHERE query = %s", (query,))
+                            local.local_db.commit()
         except Exception as e:
-            print(f"Error syncing failed queries: {e}")
-            # Error syncing failed queries: No result set to fetch from.
+            print(f"\033[91mError syncing failed queries: {e}\033[0m")
     else:
-        print("Distant database not connected")
+        print("\033[91mDistant database not connected\033[0m")
         # Create thread to check on the database
         if thread_active == 0:
             connection_thread = threading.Thread(target=connect_to_remote_db)
             connection_thread.start()
-        return 0
+        return
